Feature: speech component moved into input, display speech in text field, handleVoiceOnChange now
